refactor checkpoint 1
