+++ src/dirac_generic.h
@@ -22,6 +22,9 @@
 #ifndef DIRAC_PRECISION_HEADER
   #define DIRAC_PRECISION_HEADER
 
+#include "main_pre_def_PRECISION.h"
+#include "clifford.h"
+
   struct Thread;
   
   void gamma5_PRECISION( vector_PRECISION eta, vector_PRECISION phi, level_struct *l, struct Thread *threading );
