+++ src/global_struct.c
@@ -1,0 +1,1 @@
+global_struct g;
