+++ src/gpu/cuda_vectors_generic.h
@@ -1,0 +1,9 @@
+#ifndef CUDA_VECTORS_GENERIC_H
+#define CUDA_VECTORS_GENERIC_H
+
+#include "cuda_complex.h"
+
+typedef cu_cmplx_PRECISION* cuda_vector_PRECISION;
+typedef cu_cmplx_PRECISION* cuda_config_PRECISION;
+
+#endif //CUDA_VECTORS_GENERIC_H
