+++ src/main_pre_def_generic.h
@@ -22,22 +22,13 @@
 #ifndef MAIN_PRE_DEF_PRECISION_HEADER
   #define MAIN_PRE_DEF_PRECISION_HEADER
 
-  typedef PRECISION _Complex complex_PRECISION;
-  typedef PRECISION _Complex *config_PRECISION;
-  typedef PRECISION _Complex *vector_PRECISION;
+// FIXME temporary includes
 #ifdef CUDA_OPT
-  // CUDA-only typedefs  ---->  cuda vectors
-  typedef cu_cmplx_PRECISION* cuda_vector_PRECISION;
-  typedef cu_cmplx_PRECISION* cuda_config_PRECISION;
+#include "cuda_dirac_PRECISION.h"
 #endif
 
   struct Thread;
   struct level_struct;
-
-#ifdef CUDA_OPT
-  extern __constant__ cu_cmplx_PRECISION gamma_info_vals_PRECISION[16];
-  extern __constant__ int gamma_info_coo_PRECISION[16];
-#endif
 
   typedef struct {
     int length[8], *boundary_table[8], max_length[4],
@@ -79,17 +70,6 @@
     OPERATOR_TYPE_PRECISION *oe_clover_vectorized;
   } operator_PRECISION_struct;
 
-#ifdef CUDA_OPT
-  typedef struct {
-    cu_config_PRECISION *oe_clover_vectorized;
-    int *neighbor_table;
-    cu_config_PRECISION *D;
-    cu_cmplx_PRECISION *Dgpu[16];
-    int nr_elems_Dgpu[16];
-    cu_cmplx_PRECISION *clover_gpustorg;
-    cu_cmplx_PRECISION *oe_clover_gpustorg;
-  } operator_PRECISION_struct_on_gpu;
-#endif
   
   typedef struct {
     vector_PRECISION x, b, r, w, *V, *Z;
@@ -114,33 +94,8 @@
   } gmres_PRECISION_struct;
 
 #ifdef CUDA_OPT
-  // CUDA structs:
-  //	cuda_schwarz_PRECISION_struct:
-  //		the elements of this struct will be accessed from the CPU, but their content
-  //		are pointers pointing to GPU-data
-  //	schwarz_PRECISION_struct_on_gpu:
-  //		the elements of this struct will be accessed from within the GPU !
-  typedef struct {
-    cuda_vector_PRECISION buf1, buf2, buf3, buf4, buf5, buf6;
-    int **DD_blocks_in_comms, **DD_blocks_notin_comms, **DD_blocks;
-    block_struct* block;
-    cuda_vector_PRECISION local_minres_buffer[3];
-  } cuda_schwarz_PRECISION_struct;
-  typedef struct {
-    cu_cmplx_PRECISION* oe_buf[4];
-    cu_config_PRECISION *oe_clover_vectorized;
-    operator_PRECISION_struct_on_gpu op;
-    int num_block_even_sites, num_block_odd_sites;
-    int block_vector_size;
-    int *oe_index[4];
-    int *index[4];
-    int dir_length_even[4], dir_length_odd[4];
-    int dir_length[4];
-    int block_boundary_length[9];
-    cu_cmplx_PRECISION gamma_info_vals[16];
-    int gamma_info_coo[16];
-    cu_cmplx_PRECISION *alphas;
-  } schwarz_PRECISION_struct_on_gpu;
+//FIXME temporary inlcude
+# include "cuda_schwarz_PRECISION.h"
 #endif
 
   typedef struct {
