+++ test/test_sse_dirac_double.c
@@ -1,2 +1,6 @@
 #include <criterion/criterion.h>
-#
+#include "dirac_double.h"
+
+Test(dirac, can_include) {
+    
+}
